requirements.txt
streamlit
python-dotenv
sqlalchemy
pg8000
fastapi
requests
pydantic
seaborn
matplotlib